Refactor data uses block
